requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer backbone
timm>=0.9.0

# Data handling & utilities
numpy>=1.24.0
Pillow>=9.0.0
scipy>=1.10.0
opencv-python>=4.7.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Config & serialization
pyyaml>=6.0

# Notebook
jupyter>=1.0.0
notebook>=7.0.0

# Optional: faster attention (if using Flash Attention)
# flash-attn>=2.0.0  # requires CUDA 11.6+ and Ampere GPU

# Development
pytest>=7.0.0
black>=23.0.0
